+++ SonicMania/Objects/SSZ/HotaruMKII.c
@@ -247,14 +247,14 @@ void HotaruMKII_State_CheckPlayerInRange(void)
                 self->position.y       = RSDK.Sin256(self->origin) << 17;
                 RSDKScreenInfo *screen = &ScreenInfo[screenID];
 
-                if (self->position.x > (screen->size.x & 0xFFFFFFFE) << 15)
-                    self->position.x = (screen->size.x & 0xFFFFFFFE) << 15;
+                if (self->position.x > (int32)(screen->size.x & 0xFFFFFFFE) << 15)
+                    self->position.x = (int32)(screen->size.x & 0xFFFFFFFE) << 15;
 
                 if (self->position.x < -((int32)(screen->size.x & 0xFFFFFFFE) << 15))
                     self->position.x = -((int32)(screen->size.x & 0xFFFFFFFE) << 15);
 
-                if (self->position.y > (screen->size.y & 0xFFFFFFFE) << 15)
-                    self->position.y = (screen->size.y & 0xFFFFFFFE) << 15;
+                if (self->position.y > (int32)(screen->size.y & 0xFFFFFFFE) << 15)
+                    self->position.y = (int32)(screen->size.y & 0xFFFFFFFE) << 15;
 
                 if (self->position.y < -((int32)(screen->size.y & 0xFFFFFFFE) << 15))
                     self->position.y = -((int32)(screen->size.y & 0xFFFFFFFE) << 15);
